+++ angular_router/CHANGELOG.md
@@ -1,3 +1,8 @@
+### Bug fixes
+
+*   Navigation will no longer succeed for an empty path if it doesn't match a
+    route.
+
 ## 2.0.0-alpha+13
 
 ### New features